[#40] Deleted shared component in players folder, moved card comp to details folder
