+++ src/components/app.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
 
 import Header from './header';
-import Filters from './players/filters';
 import StandingsPage from './standings';
 import GamesPage from './games';
 import PlayersPage from './players';
@@ -14,7 +13,6 @@ const App = () => {
     <div className='App'>
       <BrowserRouter basename={process.env.PUBLIC_URL}>
         <Header />
-        <Filters />
         <div id='container'>
             <Route path='/' exact component={PlayersPage} />
             <Route path='/standings' component={StandingsPage} />
