+++ src/components/players/details/card/index.js
