+++ src/components/players/details/card/index.module.scss
