+++ src/components/players/filters/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 import styles from './index.module.scss';
 
 const Filters = () => {
-    return <div className={styles.filters} />;
+    return <div className={styles.filters}/>;
 };
 
 export default Filters;
