+++ src/components/players/index.js
@@ -2,14 +2,18 @@ import React from 'react';
 
 import List from './list';
 import Details from './details';
+import Filters from './filters';
 
 import styles from './index.module.scss';
 
 const PlayersPage = () => {
     return (
-        <div className={styles.playersContainer}>
-            <List />
-            <Details />
+        <div className={styles.container}>
+            <Filters />
+            <div className={styles.gallery}>
+                <List />
+                <Details />  
+            </div>
         </div>
     );
 };
