+++ src/components/players/index.module.scss
@@ -1,6 +1,12 @@
 @import 'src/scss/variables.scss';
 
-.playersContainer {
+.container {
     display: flex;
+    flex-direction: column;
     box-shadow: -1px 0px 1px $grey3;
+}
+
+.gallery {
+    display: flex;
+    flex-direction: row;
 }
